Convert Petition to stateless component
