+++ src/containers/sign-petition.js
@@ -4,7 +4,8 @@ import PropTypes from 'prop-types'
 import 'whatwg-fetch'
 import { connect } from 'react-redux'
 
-import Petition from '../components/petition.js'
+import Petition from 'LegacyTheme/petition'
+import { thanksLoader } from '../loaders/petition.js'
 import { actions as petitionActions } from '../actions/petitionActions.js'
 
 class SignPetition extends React.Component {
@@ -14,18 +15,32 @@ class SignPetition extends React.Component {
     dispatch(petitionActions.loadPetition(params.petition_slug))
   }
 
+  componentDidMount() {
+    // Lazy-load thanks page component
+    thanksLoader()
+  }
+
   render() {
-    if (!this.props.petition) {
+    const p = this.props.petition
+    if (!p) {
       return (
         <div>
         </div>
       )
     }
+    const creator = ((p._embedded && p._embedded.creator) || { name: p.contact_name })
+    const petitionBy = creator.name + (creator.organization
+      ? `, ${creator.organization}`
+      : '')
+    const outOfDate = (p.tags && p.tags.filter(t => t.name === 'possibly_out_of_date').length)
+
     return (
       <div className='moveon-petitions sign'>
         <Petition
           petition={this.props.petition}
           query={this.props.location.query}
+          petitionBy={petitionBy}
+          outOfDate={outOfDate}
         />
       </div>
     )
